Changes to nginx @maintenance handler
previous configuration would not display maintenance page if omero.web
wasn't running.  This was likely due to the request being passed to the
@proxy_to_app named location and the configuration not being inherited
into that stanza.

This commit moves the error_page directive up to the server container
where possible or duplicates the directive into the named location
directive otherwise.

Additionally the return code is set to 502 to avoid situations where
returning the maintenance page causes a 200 return code and confusing
monitoring systems.  (502 is always 502 with the current configuration.